Retours Design / UX / Wording     - correction apportée, en dessous de correction demandée pour les petites cases.     - Coquille dans le texte : A pris connaissance des modaCheckReadtés de traitement de mes données personnelles. (cf. capture sur les consentements)     - Ne pas faire apparaître le traitement des données si le volontaire à + de 15 ans ( ex sur ce profil)     - amélioration du champs Date pour que le datepicker s'ouvre dès qu'on click dedans.     - Field Select, refermer le select lorsqu'on choisit une valeur
