+++ admin/src/scenes/phase0/components/CorrectedRequest.js
@@ -28,11 +28,13 @@ export default function CorrectedRequest({ correctionRequest, className, reasons
   }
 
   return (
-    <div className={`relative mt-[8px] flex items-center ${className}`}>
-      <div className="bg-[#F97316] px-[10px] py-[3px] text-[#FFFFFF] text-[12px] rounded-[100px] cursor-pointer" onClick={toggleCorrectionView}>
+    <div className={`relative mt-[8px] block items-center ${className}`}>
+      <div
+        className="bg-[#F97316] px-[10px] py-[3px] text-[#FFFFFF] text-[12px] rounded-[100px] cursor-pointer inline-block mr-[8px] mb-[6px] whitespace-nowrap"
+        onClick={toggleCorrectionView}>
         Correction demandée...
       </div>
-      <div className="ml-[8px] bg-[#14B8A6] px-[10px] py-[3px] text-[#FFFFFF] text-[12px] rounded-[100px]">Correction apportée</div>
+      <div className="bg-[#14B8A6] px-[10px] py-[3px] text-[#FFFFFF] text-[12px] rounded-[100px] inline-block whitespace-nowrap">Correction apportée</div>
       {correctionView && (
         <div className="z-10 bg-white rounded-[8px] pt-[24px] pb-[8px] px-[50px] shadow-[0px_2px_26px_1px_rgba(0,0,0,0.17)] absolute left-[0px] right-[0px] top-[calc(100%+8px)]">
           <div className="text-[14px] leading-[20px] font-bold text-[#242526] mb-[3px] text-center">Ma demande de correction</div>
