+++ admin/src/scenes/phase0/components/Field.js
@@ -1,8 +1,8 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useState } from "react";
 import PencilAlt from "../../../assets/icons/PencilAlt";
 import CorrectionRequest from "./CorrectionRequest";
 import CorrectedRequest from "./CorrectedRequest";
-import ChevronDown from "../../../assets/icons/ChevronDown";
+import SimpleSelect from "./SimpleSelect";
 
 /**
  * mode  could be "correction|edition|readonly" (default readonly)
@@ -27,8 +27,6 @@ export default function Field({
   const [opened, setOpened] = useState(false);
   const [hasValidRequest, setHasValidRequest] = useState(false);
   const [requestButtonClass, setRequestButtonClass] = useState("");
-  const [selectOptionsOpened, setSelectOptionsOpened] = useState(false);
-  const selectOptionsRef = useRef();
 
   // const [editable, setEditable] = useState(false);
 
@@ -53,26 +51,6 @@ export default function Field({
       } hover:bg-[#F97316]`,
     );
   }, [mouseIn, hasValidRequest]);
-
-  useEffect(() => {
-    function handleClickOutside(e) {
-      if (selectOptionsRef.current) {
-        let target = e.target;
-        while (target) {
-          if (target === selectOptionsRef.current) {
-            return;
-          }
-          target = target.parentNode;
-        }
-        setSelectOptionsOpened(false);
-      }
-    }
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
 
   function startRequest() {
     if (group === null || group === undefined) {
@@ -99,14 +77,6 @@ export default function Field({
     editable = false;
   }
 
-  function toggleSelectOptions() {
-    setSelectOptionsOpened(!selectOptionsOpened);
-  }
-
-  function selectOption(opt) {
-    onChange && onChange(opt);
-  }
-
   return (
     <div className={className}>
       <div
@@ -117,23 +87,7 @@ export default function Field({
         {label && <label className="font-normal text-[12px] leading-[16px] text-[#6B7280]">{label}</label>}
         {mode === "edition" && editable ? (
           <>
-            {type === "select" && (
-              <div ref={selectOptionsRef}>
-                <div className="flex items-center justify-between cursor-pointer p-[5px] bg-gray-50" onClick={toggleSelectOptions}>
-                  <div className="font-normal text-[14px] leading-[20px] text-[#1F2937]">{transformer ? transformer(value) : value}</div>
-                  <ChevronDown className="text-[#1F2937] ml-[8px]" />
-                </div>
-                {selectOptionsOpened && (
-                  <div className="absolute z-10 mt-[-1] left-[0px] right-[0px] border-[#E5E7EB] border-[1px] rounded-[6px] bg-white text-[#1F2937] shadow-[0px_8px_16px_-3px_rgba(0,0,0,0.05)]">
-                    {options.map((opt) => (
-                      <div className="px-[10px] py-[5px] hover:bg-[#E5E7EB] cursor-pointer" key={opt.value} onClick={() => selectOption(opt.value)}>
-                        {opt.label}
-                      </div>
-                    ))}
-                  </div>
-                )}
-              </div>
-            )}
+            {type === "select" && <SimpleSelect value={value} transformer={transformer} options={options} onChange={onChange} />}
             {type === "text" && <input type="text" value={value} onChange={(e) => onChange(e.target.value)} className="block p-[5px] bg-gray-50 w-[100%]" />}
           </>
         ) : (
