+++ admin/src/scenes/phase0/components/FieldsGroup.js
@@ -53,7 +53,15 @@ export function FieldsGroup({
       {mode === "edition" && type === "date" ? (
         <div className={`relative bg-white py-[9px] px-[13px] border-[#D1D5DB] border-[1px] rounded-[6px] ${className}`}>
           {title && <label className="font-normal text-[12px] leading-[16px] text-[#6B7280]">{title}</label>}
-          <input type="date" value={dayjs(value).locale("fr").format("YYYY-MM-DD")} onChange={(e) => onChange(e.target.value)} className="block p-[5px] bg-gray-50 w-[100%]" />
+          <input
+            type="date"
+            value={dayjs(value).locale("fr").format("YYYY-MM-DD")}
+            onChange={(e) => onChange(e.target.value)}
+            onClick={(e) => {
+              if (e.target?.showPicker) e.target.showPicker();
+            }}
+            className="block p-[5px] bg-gray-50 w-[100%] cursor-pointer"
+          />
           {correctionRequest && correctionRequest.status === "CORRECTED" && <CorrectedRequest correctionRequest={correctionRequest} />}
         </div>
       ) : (
